以service形式，配置函数工具文档
