+++ javascript-tools/angular.json
@@ -32,9 +32,7 @@
         "@schematics/angular:pipe": {
           "spec": false
         },
-        "@schematics/angular:service": {
-          "spec": false
-        }
+
       },
       "root": "",
       "sourceRoot": "src",
