+++ javascript-tools/src/app/components/monaco-editor/monaco-editor.component.less
@@ -1,5 +1,4 @@
 #editor-container {
-  width: 800px;
-  height: 600px;
-  border: 1px solid grey
+  width: 100%;
+  height: 100%;
 }
