+++ javascript-tools/src/app/components/monaco-editor/monaco-editor.component.ts
@@ -17,7 +17,7 @@ let loadPromise: Promise<void>;
 
 @Component({
   selector: 'app-monaco-editor',
-  templateUrl:'./monaco-editor.component.html',
+  templateUrl: './monaco-editor.component.html',
   styleUrls: ['./monaco-editor.component.less'],
   providers: [
     {
@@ -29,19 +29,13 @@ let loadPromise: Promise<void>;
 })
 export class MonacoEditorComponent implements AfterViewInit, OnChanges {
   @ViewChild('editorContainer') _editorContainer: any;
-
   @Input() code = '';
   @Output() codeChange = new EventEmitter<String>();
 
-  // Holds instance of the current code editor
   // @ts-ignore
   codeEditorInstance: monaco.editor.IStandaloneCodeEditor;
-  constructor(
-    private zone: NgZone,
-    private configService: MonacoEditorConfigService
-  ) {}
+  constructor(private configService: MonacoEditorConfigService) {}
 
-  // supports two-way binding
   ngOnChanges() {
     if (this.codeEditorInstance) {
       this.codeEditorInstance.setValue(this.code);
@@ -50,7 +44,6 @@ export class MonacoEditorComponent implements AfterViewInit, OnChanges {
 
   ngAfterViewInit() {
     if (loadedMonaco) {
-      // Wait until monaco editor is available
       loadPromise.then(() => {
         this.initMonaco();
       });
@@ -62,23 +55,19 @@ export class MonacoEditorComponent implements AfterViewInit, OnChanges {
           return;
         }
         const onAmdLoader: any = () => {
-          // Load monaco
           (<any>window).require.config({ paths: { vs: 'assets/monaco/vs' } });
-
           (<any>window).require(['vs/editor/editor.main'], () => {
             this.initMonaco();
             resolve();
           });
         };
-        console.log(onAmdLoader);
-        // Load AMD loader if necessary
+
         if (!(<any>window).require) {
           const loaderScript: HTMLScriptElement =
             document.createElement('script');
           loaderScript.type = 'text/javascript';
           loaderScript.src = 'assets/monaco/vs/loader.js';
           loaderScript.addEventListener('load', onAmdLoader);
-          console.log(loaderScript);
           document.body.appendChild(loaderScript);
         } else {
           onAmdLoader();
@@ -88,7 +77,6 @@ export class MonacoEditorComponent implements AfterViewInit, OnChanges {
   }
 
   initMonaco(): void {
-    // configure the monaco editor to understand custom language - customLang
     monaco.languages.register(this.configService.getCustomLangExtensionPoint());
     monaco.languages.setMonarchTokensProvider(
       'CustomLang',
@@ -97,7 +85,7 @@ export class MonacoEditorComponent implements AfterViewInit, OnChanges {
     monaco.editor.defineTheme(
       'customLangTheme',
       this.configService.getCustomLangTheme()
-    ); // add your custom theme here
+    );
 
     this.codeEditorInstance = monaco.editor.create(
       this._editorContainer.nativeElement,
@@ -108,7 +96,6 @@ export class MonacoEditorComponent implements AfterViewInit, OnChanges {
       }
     );
 
-    // To support two-way binding of the code
     // @ts-ignore
     this.codeEditorInstance.getModel().onDidChangeContent((e: any) => {
       this.codeChange.emit(this.codeEditorInstance.getValue());
