+++ javascript-tools/src/app/pages/util-one/util-one.component.html
@@ -1,6 +1,2 @@
 
 <app-monaco-editor [code]='initCode'></app-monaco-editor>
-<div>
-    <button (click)="saveCode()"> Save </button>
-    <button (click)="clearCode()"> Clear </button>
-</div>
